refactor: удалены лишние миддлвары, настройка сессии перенеса в миддлвар
