+++ src/server/app.ts
@@ -6,35 +6,29 @@ import {
   loggerMiddleware,
   renderMiddleware,
   staticMiddleware,
+  createSessionMiddlewares,
 } from '@/server/middlewares';
 import { apiRoute, healthChecksRoute } from '@/server/routes';
 import { initializeDB } from '@/database';
-import { createSessionMiddlewares } from '@/server/config/session';
-
-const dbInitializeResultPromise = initializeDB();
 
 export const createApp = async () => {
-  const dbInitializeResult = await dbInitializeResultPromise;
+  const dbInitializeResult = await initializeDB();
   if (dbInitializeResult instanceof Error) {
     return null;
   }
   const { sequelizeSessionStore } = dbInitializeResult;
 
-  return (
-    express()
-      .disable('x-powered-by')
-      .enable('trust proxy')
-      .use(staticMiddleware)
-      .use(createSessionMiddlewares({ store: sequelizeSessionStore }))
-      .use(cookieParser())
-      .use(express.json())
-      .use(express.urlencoded({ extended: false }))
-      // .use('/yandex-api', yandexApiRoute)
-      .use(loggerMiddleware)
-      .use('/hc', healthChecksRoute)
-      // .use(getYandexUserMiddleware)
-      .use('/api', apiRoute)
-      .get('*', renderMiddleware)
-      .use(errorHandlerMiddleware)
-  );
+  return express()
+    .disable('x-powered-by')
+    .enable('trust proxy')
+    .use(staticMiddleware)
+    .use(createSessionMiddlewares({ store: sequelizeSessionStore }))
+    .use(cookieParser())
+    .use(express.json())
+    .use(express.urlencoded({ extended: false }))
+    .use(loggerMiddleware)
+    .use('/hc', healthChecksRoute)
+    .use('/api', apiRoute)
+    .get('*', renderMiddleware)
+    .use(errorHandlerMiddleware);
 };
