+++ src/server/config/index.ts
@@ -5,4 +5,3 @@ export const port = normalizePort(process.env.PORT || '3000');
 export { env } from '@/config';
 export { default as projectPaths } from './projectPaths';
 export * as oauth2 from './oauth2';
-export * as session from './session';
