+++ src/server/middlewares/index.ts
@@ -1,6 +1,5 @@
 export { default as renderMiddleware } from './render';
 export { default as loggerMiddleware } from './logger';
 export { default as errorHandlerMiddleware } from './errorHandler';
-export { default as getYandexUserMiddleware } from './getYandexUser';
-export { default as checkYandexUserMiddleware } from './checkYandexUser';
 export { default as staticMiddleware } from './static';
+export { createSessionMiddlewares } from './session';
